+++ app/models/releaf/node.rb
@@ -1,7 +1,5 @@
 module Releaf
   class Node < ActiveRecord::Base
-    COMMON_FIELD_NAME_PREFIX = 'data_'
-    COMMON_FIELDS_SCHEMA_FILENAME = Rails.root.to_s + '/config/common_fields.yml'
 
     ## for some reason many things will fail when Releaf:Node is abstract class
     ## TODO find out why
@@ -10,7 +8,6 @@ module Releaf
     acts_as_nested_set order_column: :item_position
     acts_as_list scope: :parent_id, column: :item_position
 
-    serialize :data, Hash
     default_scope { order(:item_position) }
 
     validates_presence_of :name, :slug, :content_type
@@ -50,67 +47,6 @@ module Releaf
 
     def to_s
       name
-    end
-
-    def method_missing(method, *args, &block)
-      if self.class.column_names.include?('data')
-        case method.to_s
-        when /^(#{COMMON_FIELD_NAME_PREFIX}(.+))=$/ then
-          if common_field_names.include? $1
-            return common_field_setter($2, args[0])
-          end
-
-        when /^(#{COMMON_FIELD_NAME_PREFIX}(.+))$/ then
-          if common_field_names.include? $1
-            return common_field_getter($2)
-          end
-        end
-      end
-
-      return super(method, *args, &block)
-    end
-
-    def respond_to? method_id, include_private = false
-      # this is needed for mass asignment (update_attributes) to work with common_fields
-      rt = super(method_id, include_private)
-      return true if rt
-      return common_field_names.include?(method_id.to_s.sub(/=$/, ''))
-    end
-
-    def common_fields_schema
-      full_schema = common_fields_full_schema
-      full_schema.delete_if do |field|
-        common_fields_keep_field?(field) == false
-      end
-      return full_schema
-    end
-
-
-    def common_field_names
-      common_fields_schema.map { |f| "#{COMMON_FIELD_NAME_PREFIX}#{f['field_name']}" }
-    end
-
-    def common_field_field_type(field_name)
-      common_field_options(field_name.sub(/^#{COMMON_FIELD_NAME_PREFIX}/, ''))['field_type']
-    end
-
-    def self.load_common_fields_schema
-      cfschema = if File.exists?(COMMON_FIELDS_SCHEMA_FILENAME)
-                   YAML::load_file(COMMON_FIELDS_SCHEMA_FILENAME)
-                 else
-                   []
-                 end
-
-      raise "common_fields common_fields_schema is not an array" unless cfschema.is_a? Array
-      cfschema.each_with_index do |field, i|
-        raise "common_fields common_fields_schema contains non-hash element in root node"       unless field.is_a? Hash
-        raise "field_name not defined for field ##{i}"                            unless field.has_key?('field_name')
-        raise "field_name must be string"                                         unless field['field_name'].is_a? String
-        raise "field_type not defined for #{field['field_name']} field"           unless field.has_key?('field_type')
-        raise "apply_to not defined for #{field['field_name']} field"             unless field.has_key?('apply_to')
-      end
-
-      return cfschema
     end
 
     def destroy
@@ -228,47 +164,6 @@ module Releaf
 
     def update_settings_timestamp
       Settings['nodes.updated_at'] = Time.now
-    end
-
-    def common_fields_full_schema
-      self.class.load_common_fields_schema.dup
-    end
-
-    def common_fields_keep_helper field_options, mode, match_value, default_value=false
-      if ['*', content_type].include? field_options[mode]
-        return match_value
-      elsif field_options[mode].is_a?(Array) && field_options[mode].include?(content_type)
-        return match_value
-      else
-        return default_value
-      end
-    end
-
-    def common_fields_keep_field? field_options
-      keep = common_fields_keep_helper(field_options, 'apply_to', true, false)
-      keep = common_fields_keep_helper(field_options, 'deny_for', false, keep) if keep
-      if keep && field_options.has_key?('levels')
-        keep = false unless field_options['levels'].include? depth
-      end
-      return keep
-    end
-
-    def common_field_setter(key, value)
-      self.data[key] = value
-    end
-
-    def common_field_getter(key)
-      self.data.fetch(key, common_field_default_value(key))
-    end
-
-    def common_field_options(key)
-      common_fields_schema.each do |field|
-        return field if field['field_name'] == key
-      end
-    end
-
-    def common_field_default_value(key)
-      common_field_options(key)['default']
     end
 
   end